refactor: move context definition tp types
