+++ frontend/svelte/src/tests/mocks/transaction.store.mock.ts
@@ -1,5 +1,5 @@
 import { writable } from "svelte/store";
-import type { TransactionStore } from "../../lib/stores/transaction.store";
+import type { TransactionStore } from "../../lib/types/transaction.context";
 
 export const mockTransactionStore = writable<TransactionStore>({
   selectedAccount: undefined,
